db_helper/db_python_app_template/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
pyodbc==5.0.1
pydantic==2.4.2
flask_pydantic